Day 6 part 2 done. ☀️ 🎉😰
I couldn't figure out for my life what lead to a few extra positive
loops earlier. :/ Not excluding the starting position definitely wasn't
the issue since that would be a single additional loop.

Starting a new walk with a single obstacle placed along the original
path should be the same as placing an obstacle immediately in front
of the guard's's path (excluding out of bound and `#`), so I probably
made a silly mistake somewhere. :/