Implement `solvePart2()`.
The algorithm relies on there are only large amount of unique interconnect groups when the are relatively small, and that number would rapidly decrease as their sizes grow.

Checking every connection on every single group is still kind of acceptable (~3 seconds on an M3). I'm too stupid to figure out a better way to do this.